Move the caching test to the common trait
